ui manages a 'clock' channel
An int value is received on each Device.ticker message. The value is the
total message count since it began playing.

Device.ticker interval increased by 24*. This aligns with midi spec: 24
clock messages per quarter note. This will enable parts to use smaller
divisions and to be able to sync with other devices.